+++ src/test/groovy/com/moowork/gradle/grunt/GruntTaskTest.groovy
@@ -1,0 +1,20 @@
+package com.moowork.gradle.grunt
+
+import org.gradle.api.GradleException
+
+class GruntTaskTest
+    extends AbstractTaskTest
+{
+    def "grunt not installed"()
+    {
+        given:
+        def task = this.project.tasks.create( 'simple', GruntTask )
+
+        when:
+        this.project.evaluate()
+        task.exec()
+
+        then:
+        thrown( GradleException )
+    }
+}