+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version 1.0 *(not released)*
 ----------------------------
 
 * Move npm and yarn classes into separate package (#158)
+* Move grunt plugin code to this plugin (#159)
 * ...
 
 Version 0.14 *(2016-11-29)*